pyproject.toml
[project]
name = "course-scout"
version = "0.2.0"
description = "Telegram art channel scanner and digest generator"
readme = "README.md"
authors = [{ name = "Adam", email = "adam@example.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx[socks]>=0.28.1",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "tenacity>=8.2.3",
    "telethon>=1.39.0",
    "typer>=0.15.1",
    "uvicorn>=0.40.0",
    "weasyprint>=61.0",
    "markdown>=3.5.2",
    "fpdf2>=2.8.5",
    "jsonref>=1.1.0",
    "markdown2>=2.5.4",
    "markdown-pdf>=1.10",
    "mcp[cli]>=1.25.0",
    "PyYAML>=6.0.1",
    "claude-agent-sdk>=0.1.50",
    "openai>=2.30.0",
]

[project.scripts]
course-scout = "course_scout.interfaces.cli.main:app"
course-scout-worker = "course_scout.application.worker:main"
course-scout-sse = "course_scout.interfaces.api.sse:main"

[dependency-groups]
dev = [
    "import-linter>=2.9",
    "pyright>=1.1.408",
    "pytest-cov>=7.0.0",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
    "pytest-asyncio>=1.3.0",
    "rust-just>=1.46.0",
]

[tool.importlinter]
root_package = "course_scout"

[[tool.importlinter.contracts]]
name = "Domain Driven Architecture Layers"
type = "layers"
layers = [
    "interfaces",
    "application",
    "infrastructure",
    "domain",
]
containers = ["course_scout"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# E: pycodestyle errors
# F: Pyflakes
# I: isort
# B: flake8-bugbear
# C4: flake8-comprehensions
# UP: pyupgrade
# W: pycodestyle warnings
# C90: mccabe complexity
# D: pydocstyle (documentation)
# FIX: flake8-fixme (TODO detection)
select = ["E", "F", "I", "B", "C4", "UP", "W", "TID252", "C90", "D", "FIX"]
ignore = ["D100", "D101", "D102", "D103", "D104"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
include = ["src"]
typeCheckingMode = "standard"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=src/course_scout --cov-report=term-missing"

[tool.coverage.report]
fail_under = 80
show_missing = true
omit = ["tests/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
